wineproject/requirements.txt
django-filter
django-widget-tweaks
pymysql
